feat(translations): KOR-60: Import default translations in auto-deploys (#8445)
