+++ .github/workflows/ci.yml
@@ -392,20 +392,6 @@ jobs:
         with:
           config: .typos.toml
 
-  translations:
-    name: Check for duplicates in translation stringIds
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: actions/setup-node@v4
-        with:
-          node-version-file: .node-version
-          cache: npm
-      - uses: taiki-e/install-action@v2
-        with:
-          tool: ripgrep
-      - run: .github/scripts/scrape-translations.sh
-
   # Dummy job to have a stable name for PR requirements
   tests-pass:
     if: always() # always run even if dependencies fail
@@ -422,7 +408,6 @@ jobs:
       - dbt-model
       - dbt-todos
       - typos
-      - translations
     runs-on: ubuntu-latest
     steps:
       - uses: re-actors/alls-green@release/v1
