--- .github/workflows/diff-translations.yml
@@ -1,122 +1,0 @@
-name: Diff Translations
-
-on:
-  push:
-    tags:
-      - 'v*'
-
-env:
-  RG_VERSION: 14.1.0
-
-jobs:
-  run:
-    name: Diff Translations
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write
-      id-token: write
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          token: ${{ secrets.TAMANU_RELEASE_PAT }}
-          fetch-depth: 0
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version-file: .node-version
-          cache: npm
-      - uses: taiki-e/install-action@v2
-        with:
-          tool: ripgrep
-
-      - name: get current version
-        id: currentv
-        run: |
-          version=$(jq -r .version package.json)
-          echo "current_version=$version" >> $GITHUB_OUTPUT
-
-      - name: get this versions translations
-        run: |
-          translations=$(.github/scripts/scrape-translations.sh)
-          if [ $? -eq 0 ]; then
-           echo "$translations" > current-translations.csv
-          else
-            echo "Scrape translations script failed with exit code $?"
-          fi
-
-      - name: Configure AWS Credentials (upload)
-        uses: aws-actions/configure-aws-credentials@v4
-        with:
-          aws-region: ap-southeast-2
-          role-to-assume: arn:aws:iam::143295493206:role/gha-tamanu-translations-upload
-          role-session-name: GHA@Tamanu=DiffTranslations
-
-      - name: Upload translations to S3
-        env:
-          version: ${{ steps.currentv.outputs.current_version }}
-        run: |
-          aws s3 --no-progress cp current-translations.csv s3://bes-tamanu-translations/v$version.csv
-
-      - name: Post to meta
-        continue-on-error: true
-        uses: ./.github/actions/meta-api
-        with:
-          url: ${{ vars.META_URL }}
-          crt: ${{ secrets.META_CERT }}
-          key: ${{ secrets.META_KEY }}
-          api: /artifacts/${{ steps.currentv.outputs.current_version }}/translations/any
-          arg: --data-binary 'https://translations.ops.tamanu.io/v${{ steps.currentv.outputs.current_version }}.csv'
-
-      - name: Post to meta
-        continue-on-error: true
-        uses: ./.github/actions/meta-api
-        with:
-          url: ${{ vars.META_URL }}
-          crt: ${{ secrets.META_CERT }}
-          key: ${{ secrets.META_KEY }}
-          api: /artifacts/${{ steps.currentv.outputs.current_version }}/translations/any
-          arg: --data-binary 'https://translations.ops.tamanu.io/v${{ steps.currentv.outputs.current_version }}.csv'
-
-      - name: get closest previous version
-        id: previousv
-        run: |
-          set -euxo pipefail
-          git fetch --tags --force
-          git tag --list \
-            | grep -E '^v[0-9]+[.][0-9]+[.][0-9]+$' \
-            | cut -c2- \
-            | sort --version-sort \
-            | grep -B 1 ${{steps.currentv.outputs.current_version}} \
-            | head -n 1 \
-            | sed 's/^/previous_version=/' \
-            | tee -a $GITHUB_OUTPUT
-
-      - name: Download previous versions translations from s3
-        continue-on-error: true
-        env:
-          pversion: ${{ steps.previousv.outputs.previous_version }}
-        run: |
-          aws s3 --no-progress cp s3://bes-tamanu-translations/v$pversion.csv previous-translations.csv
-
-      - name: Download previous versions translations from s3
-        continue-on-error: true
-        env:
-          pversion: ${{ steps.previousv.outputs.previous_version }}
-        run: |
-          aws s3 --no-progress cp s3://bes-tamanu-translations/v$pversion.csv previous-translations.csv
-
-      - name: Diff translations
-        run: |
-          cat <<EOF | tee RELEASE_NOTES.md
-          ## Translation Changes
-          *Changes compared to previous version: v${{ steps.previousv.outputs.previous_version }}*
-          ```diff
-          $(git diff --no-index previous-translations.csv current-translations.csv || true)
-          ```
-          EOF
-
-      - uses: softprops/action-gh-release@v2
-        with:
-          tag_name: ${{ github.ref_name }}
-          append_body: true
-          body_path: RELEASE_NOTES.md
