--- packages/constants/scripts/printTranslatedEnums.ts
@@ -1,3 +1,0 @@
-import { enumTranslations } from '../src/enumRegistry.js';
-
-console.log(enumTranslations.reduce((acc, [key, value]) => `"${key}","${value}"\n${acc}`, ''));
